EjbWithNr changed from int to long

+++ interfaces/src/main/java/net/sf/ahtutils/interfaces/model/with/EjbWithNr.java
@@ -2,6 +2,6 @@ package net.sf.ahtutils.interfaces.model.with;
 
 public interface EjbWithNr
 {	
-	int getNr();
-	void setNr(int nr);
+	long getNr();
+	void setNr(long nr);
 }
